+++ NEWS.md
@@ -1,5 +1,9 @@
 # swfscDAS 0.3 (in development)
 
+* `das_process` now lets the user add the DateTime, Lat, and Lon information to ? and numeric (1:8) events. This allows the user to filter the processed DAS data by date/time or coordinates, if desired.
+
+* Added `das_within_strata` for checking if individual points, such as event points or segment midpoints, are within user-provided strata
+
 
 # swfscDAS 0.2